Exclude Output for Swift package
Use .gitkeep to ensure the Output folder exists on a clean checkout (as on the GitHub Actions runner). This file need to also be present in Content otherwise it will be deleted when building the site. The downside is this file to be present in the deployed site, but that doesn’t really matter.